project name and structure refactored
