+++ src/main/kotlin/com/sentinex/sentinex_backend/SentimentApplication.kt
@@ -1,4 +1,4 @@
-package com.sentinex.sentiment_backend
+package com.sentinex.sentinex_backend
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
