+++ src/test/kotlin/com/sentinex/sentinex_backend/SentimentApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.sentinex.sentiment_backend
+package com.sentinex.sentinex_backend
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
